Introduce $logger.printMarkdown
This logic is used when printing help on the console, as well as for command suggestions.
