+++ commands/help.ts
@@ -27,7 +27,7 @@ export class HelpCommand implements ICommand {
 
 			if(this.$options.help) {
 				let help = this.$htmlHelpService.getCommandLineHelpForCommand(topic).wait();
-				this.$logger.out(help);
+				this.$logger.printMarkdown(help);
 			} else {
 				this.$htmlHelpService.openHelpForCommandInBrowser(topic).wait();
 			}
